chore(e2e): move commands to commands folder
