+++ packages/compass-e2e-tests/helpers/commands/index.ts
@@ -62,3 +62,5 @@ export * from './unhide-index';
 export * from './hide-visible-modal';
 export * from './hide-visible-toasts';
 export * from './sidebar-collection';
+export * from './read-first-document-content';
+export * from './read-stage-operators';
